requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
selenium==4.18.1
undetected-chromedriver>=3.5.0
setuptools>=65.0.0
PyPDF2==3.0.1
python-dateutil==2.8.2
cloudscraper>=1.2.71